Refactor CSS loading to initialization instead of React component
